+++ Tests/JsonPatchRfc7396.Scalar/JsonPatchRfc7396.Scalar.Application/Documents/PatchDocumentChange/PatchDocumentChangeCommandHandler.cs
@@ -41,7 +41,7 @@ namespace JsonPatchRfc7396.Scalar.Application.Documents.PatchDocumentChange
                 throw new NotFoundException($"Could not find DocumentChange '{request.Id}'");
             }
             LoadOriginalState(documentChange, request);
-            request.PatchExecutor.ApplyTo(request);
+            await request.PatchExecutor.ApplyToAsync(request, cancellationToken);
             ApplyChangesTo(request, documentChange);
 
             _documentRepository.Update(document);