+++ Tests/JsonPatchRfc7396.Scalar/JsonPatchRfc7396.Scalar.Application/Implementation/ConfigurationStoresService.cs
@@ -100,7 +100,7 @@ namespace JsonPatchRfc7396.Scalar.Application.Implementation
             }
 
             LoadOriginalState(configurationStore, dto);
-            dto.PatchExecutor.ApplyTo(dto);
+            await dto.PatchExecutor.ApplyToAsync(dto, cancellationToken);
             ApplyChangesTo(dto, configurationStore);
             return configurationStore.MapToConfigurationStoreDto(_mapper);
         }
@@ -157,7 +157,7 @@ namespace JsonPatchRfc7396.Scalar.Application.Implementation
                 throw new NotFoundException($"Could not find ConfigurationItem '{id}'");
             }
             LoadOriginalState(configurationItem, dto);
-            dto.PatchExecutor.ApplyTo(dto);
+            await dto.PatchExecutor.ApplyToAsync(dto, cancellationToken);
             ApplyChangesTo(dto, configurationItem);
             return configurationItem.MapToConfigurationConfigurationItemDto(_mapper);
         }