+++ Tests/SqlDbProject.Sdk/SqlDbProject.Sdk.SqlDatabaseProject.Sdk/accountholder/Stored Procedures/GetAccountHolderPerson.sql
@@ -1,0 +1,4 @@
+CREATE PROCEDURE [accountholder].[GetAccountHolderPerson]
+    @stakeholderId BIGINT
+AS
+SET NOCOUNT ON;