+++ Tests/SqlDbProject/SqlDbProject.SqlDatabaseProject.DNFramework/finance/Tables/AccountType.sql
@@ -1,0 +1,6 @@
+CREATE TABLE [finance].[AccountType]
+(
+    [AccountTypeId] INT NOT NULL,
+    [Description] NVARCHAR(255) NOT NULL,
+    CONSTRAINT [PK_AccountType] PRIMARY KEY CLUSTERED ([AccountTypeId] ASC)
+);