+++ Modules/Intent.Modules.EntityFrameworkCore/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 5.0.42
 
 - Fixed: NullReferenceException when LazyLoadingWithProxies is enabled but no Entity template could be found.
+- Fixed: If you model a composite entity that inherits from a non-abstract entity it should report that it is not supported.
 
 ### Version 5.0.41