+++ Modules/Intent.Modules.EntityFrameworkCore/release-notes.md
@@ -1,3 +1,8 @@
+### Version 5.0.42
+
+- Fixed: NullReferenceException when LazyLoadingWithProxies is enabled but no Entity template could be found.
+- Fixed: If you model a composite entity that inherits from a non-abstract entity it should report that it is not supported.
+
 ### Version 5.0.41
 
 - Improvement: When Domain Constraints are installed, using `Text Limits` can also be used to specify string column lengths if `Text Constraints` are not being used.